Adicionado projeto Entity Framework.

+++ src/Oo.Ddd.Bank.WebApp/Models/ContaModel.cs
@@ -2,5 +2,5 @@
 {
     public string NumeroConta { get; set; }
     public string Senha { get; set; }
-    public decimal Saldo { get; set; }
+    public double Saldo { get; set; }
 }
